fix: clean up server function to be JS only
